+++ Enseignat-Service/src/main/java/com/enseignant/response/CourResponse.java
@@ -12,7 +12,7 @@ public class CourResponse {
 	private String intitule;
 	private String dateDebut;
 	private String dateUpdate;
-	private long idModule;
+	private Integer idModule;
 	private long id_enseignant;
 	private String moduleName;
 	private String enseignant_name;
